refactor(deploy): デプロイ通知のファイルを1つに集約
